pyproject.toml
[project]
name = "test-hitalent"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Novojilov",email = "wowlass63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "docker (>=7.1.0,<8.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "ruff (>=0.12.11,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
